add BitStatusUtil
